Refactor import statements in storage and todolist components
