+++ client/app/(main)/storage/components/index.ts
@@ -1,3 +1,5 @@
+'use client'
+
 export * from './StorageSection'
 export * from './StorageHeader'
 export * from './StorageListDisplay'
