+++ client/app/(main)/todolist/components/index.ts
@@ -1,3 +1,5 @@
+'use client'
+
 export * from './CreateTodolist'
 export * from './TodolistSection'
 export * from './TodolistHeader'
