+++ client/app/(main)/todolist/hooks/index.ts
@@ -1,3 +1,5 @@
+'use client'
+
 export * from './useDragDndKit'
 export * from './useTodolist'
 export * from './useTodolistEditModal'
